starter-kits/claude-agent-sdk/requirements.txt
anthropic>=0.50.0
python-dotenv>=1.0.0
streamlit>=1.35.0
